requirements.txt
# Motor de Calculo BT - dependencias
# Instalar con: pip install -r requirements.txt

openpyxl==3.1.5      # leer y escribir archivos Excel
pytest==9.0.2        # ejecutar tests unitarios
pyinstaller==6.19.0  # generar ejecutable .exe
python-docx==1.2.0   # generar memoria explicativa DOCX
reportlab==4.4.10     # generar reporte cliente PDF
llama-index-core==0.13.6
llama-index-vector-stores-chroma==0.6.3
chromadb==1.0.20
sentence-transformers==5.1.0
pdfplumber==0.11.7
matplotlib>=3.8

